Revised add media to collections with the ability to have custom orders
Plex Smart Collections cannot have a custom order and is limited for sorting options since you cannot have the order of your choice set. Making use of non-smart collections avoids this issue and allows for unlimited number of custom ordered lists. When the script is run again the collection is not deleted therefore allowing for it to be pinned on users home screen and allows for it to be updated by changes in your list.